+++ src/main/java/com/springboot/practicebank/controller/UserController.java
@@ -29,7 +29,4 @@ public class UserController {
     public LoginResponse userLogin(@RequestBody @Valid LoginRequest loginRequest){
         return userService.userLogin(loginRequest);
     }
-
-
-
 }
