+++ src/main/java/com/springboot/practicebank/dto/LoginRequest.java
@@ -20,4 +20,5 @@ public class LoginRequest {
     private String email;
 
     private String password;
+
 }
