--- src/main/java/com/springboot/practicebank/service/UserDetailService.java
@@ -1,27 +1,0 @@
-package com.springboot.practicebank.service;
-
-import com.springboot.practicebank.repository.UserRepository;
-import lombok.AllArgsConstructor;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-import java.util.NoSuchElementException;
-
-@Service
-
-@AllArgsConstructor
-public class UserDetailService implements UserDetailsService {
-
-    private final UserRepository userRepository;
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-
-        return userRepository.findByEmail(username)
-            .orElseThrow(() -> new NoSuchElementException("User not Found: " + username));
-
-    }
-
-}
-
